// src/app/models/contact.model.ts
export class Contact {
  _id?: string;
  userId: string;
  companyId: string;
  companyPosition: string;
  stage: string;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  address: string;
  facebook: string;
  linkedin: string;
  twitter: string;
  instagram: string;
  github: string;
  youtube: string;
}